Update SyncFreedom connections and bump version
Refactored the SyncFreedomQBOConnection class to accept either credentials or a dictionary for creating the connection. Added a new class, SyncFreedomQBOConnections, to manage a list of QBOConnections. Also, implemented relevant tests for these updated features. Updated the package version to 1.0.5. This will ensure more flexibility in establishing connections and managing multiple connections.