Refactor SyncFreedom classes for more flexible connections
Expanded functionality in the SyncFreedomQBOConnection class and added the class SyncFreedomQBOConnections for managing multiple connections. This update allows for a QBOConnection to be made with either credentials or a dictionary, offering more flexibility in establishing connections. The test suite was accordingly updated to ensure functionality, and the package version was bumped to 1.0.5.