Update SyncFreedomQuickBooks class and improve connection flexibility
This commit refactors the `SyncFreedomQuickBooks` class in `client.py` allowing for a QuickBooks Online Connection to be established only with credentials, reducing complexity and improving flexibility. It also introduces a `company_id` property in the `SyncFreedomQBOConnection` class. README.md, setup.py, setup.cfg and LICENSE files were updated accordingly and `.pypirc` was added to `.gitignore`. The package version was incremented in the setup files.